+++ pallets/evercity-carbon-credits/Cargo.toml
@@ -31,10 +31,6 @@ pallet-timestamp = { default-features = false, version = '3.0.0' }
 sp-api = { default-features = false, version = '3.0.0' }
 sp-runtime = { default-features = false, version = '3.0.0' }
 sp-std = { default-features = false, version = '3.0.0' }
-# pallet-evercity-assets = { default-features = false, version = '0.1.0', git = 'https://github.com/EvercityEcosystem/evercity-assets' }
-# pallet-evercity-accounts = { default-features = false, version = '0.1.7', git = 'https://github.com/EvercityEcosystem/evercity-accounts' }
-# # add filesign with temp branch:
-# pallet-evercity-filesign = { default-features = false, version = '0.1.4', git = 'https://github.com/EvercityEcosystem/filesign'}
 pallet-evercity-bonds ={ default-features = false, version = '0.1.1', path = '../evercity-bonds' }
 pallet-evercity-assets = { default-features = false, version = '0.1.0', path = '../evercity-assets' }
 pallet-evercity-accounts = { default-features = false, version = '0.1.7', path = '../evercity-accounts'  }
