+++ pallets/evercity-carbon-credits/src/tests/mod.rs
@@ -2,6 +2,7 @@ pub mod project_tests;
 pub mod annual_report_tests;
 pub mod carbon_credits_tests;
 pub mod bond_cc_tests;
+pub mod cc_exchange_tests;
 
 pub mod helpers;
 pub mod mock;
