--- pallets/evercity-exchange/Cargo.toml
@@ -1,65 +1,0 @@
-[package]
-authors = ["Dima Bushuev diman1ghtly@gmail.com"]
-description = 'Pallet for Evercity EverUSD and Carbon Credits exchange'
-edition = '2018'
-homepage = 'https://evercity.io'
-license = 'Apache-2.0'
-name = 'pallet-evercity-exchange'
-readme = '../README.md'
-repository = 'https://github.com/EvercityEcosystem/evercity-substrate'
-version = '0.1.0'
-
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-[dependencies.serde]
-optional = true
-features = ['derive']
-version = '1.0.124'
-
-[package.metadata.docs.rs]
-targets = ['x86_64-unknown-linux-gnu']
-[dev-dependencies.serde]
-version = '1.0.119'
-
-[dependencies.codec]
-default-features = false
-features = ['derive']
-package = 'parity-scale-codec'
-version = '2.0.0'
-
-[dependencies]
-frame-support = { default-features = false, version = '3.0.0' }
-frame-system = { default-features = false, version = '3.0.0' }
-sp-api = { default-features = false, version = '3.0.0' }
-sp-runtime = { default-features = false, version = '3.0.0' }
-sp-std = { default-features = false, version = '3.0.0' }
-pallet-timestamp = { default-features = false, version = '3.0.0' }
-pallet-evercity-carbon-credits = { default-features = false, version = '0.1.13', path = '../evercity-carbon-credits' }
-pallet-evercity-assets = { default-features = false, version = '0.1.0', path = '../evercity-assets' }
-pallet-evercity-bonds ={ default-features = false, version = '0.1.1', path = '../evercity-bonds' }
-
-[dev-dependencies]
-sp-core = {  version = '3.0.0' }
-sp-io = { default-features = false, version = '3.0.0' }
-pallet-balances = { default-features = false, version = '3.0.0' }
-pallet-evercity-accounts = { default-features = false, version = '0.1.7', path = '../evercity-accounts' }
-pallet-evercity-filesign = { default-features = false, version = '0.1.4', path = '../evercity-filesign' }
-pallet-timestamp = { default-features = false, version = '3.0.0' }
-pallet-randomness-collective-flip = { default-features = false, version = '3.0.0' }
-
-[features]
-default = ['std']
-std = [
-    'serde',
-    'codec/std',
-    'sp-api/std',
-    'frame-support/std',
-    'frame-system/std',
-    'pallet-timestamp/std',
-    'pallet-evercity-carbon-credits/std',
-    'pallet-evercity-assets/std',
-    'pallet-balances/std',
-    'pallet-evercity-accounts/std',
-    'pallet-evercity-filesign/std',
-]
-
