--- pallets/evercity-exchange/makefile
@@ -1,11 +1,0 @@
-build:
-	SKIP_WASM_BUILD= cargo build --release
-
-test:
-	SKIP_WASM_BUILD= cargo test
-
-check:
-	cargo check --all --tests
-
-lint:
-	cargo clippy --all-targets
